changed field median to double for more accurate results

+++ src/main/java/com/of/scraper/dto/AverageAndMedianDTO.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class AverageAndMedianDTO {
     
     double average;
-    int median;
+    double median;
 }
